moving more out of qutip_unitary_solver to pulse_controller, and re-enabled cython pieces and cython testing as I don't want that to break yet
