+++ qiskit/providers/aer/pulse/de_solvers/qutip_unitary_solver.py
@@ -22,123 +22,9 @@ import logging
 import numpy as np
 from scipy.integrate import ode
 from scipy.linalg.blas import get_blas_funcs
-from qiskit.tools.parallel import parallel_map, CPU_COUNT
-from ..pulse0.qutip_lite.cy.spmatfuncs import cy_expect_psi_csr
-from ..pulse0.qutip_lite.cy.utilities import _cython_build_cleanup
 from ..pulse0.cy.measure import occ_probabilities, write_shots_memory
 
 dznrm2 = get_blas_funcs("znrm2", dtype=np.float64)
-
-#
-# Internal, global variables for storing references to dynamically loaded
-# cython functions
-#
-_cy_rhs_func = None
-
-def run_unitary_experiments(op_system):
-
-
-    """unitary evolution requires no seeds, so move this out of this deterministic DE
-    solving class once measurements are moved
-    """
-    # setup seeds array
-    if op_system.global_data['seed']:
-        prng = np.random.RandomState(op_system.global_data['seed'])
-    else:
-        prng = np.random.RandomState(
-            np.random.randint(np.iinfo(np.int32).max - 1))
-    for exp in op_system.experiments:
-        exp['seed'] = prng.randint(np.iinfo(np.int32).max - 1)
-
-
-    map_kwargs = {'num_processes': op_system.ode_options.num_cpus}
-
-
-    start = time.time()
-    exp_results = parallel_map(unitary_evolution,
-                               op_system.experiments,
-                               task_args=(op_system,),
-                               **map_kwargs
-                               )
-    end = time.time()
-    exp_times = (np.ones(len(op_system.experiments)) *
-                 (end - start) / len(op_system.experiments))
-
-
-    # format the data into the proper output
-    all_results = []
-    for idx_exp, exp in enumerate(op_system.experiments):
-
-        m_lev = op_system.global_data['meas_level']
-        m_ret = op_system.global_data['meas_return']
-
-        # populate the results dictionary
-        results = {'seed_simulator': exp['seed'],
-                   'shots': op_system.global_data['shots'],
-                   'status': 'DONE',
-                   'success': True,
-                   'time_taken': exp_times[idx_exp],
-                   'header': exp['header'],
-                   'meas_level': m_lev,
-                   'meas_return': m_ret,
-                   'data': {}}
-
-        memory = exp_results[idx_exp][0]
-        results['data']['statevector'] = []
-        for coef in exp_results[idx_exp][1]:
-            results['data']['statevector'].append([np.real(coef),
-                                                   np.imag(coef)])
-        results['header']['ode_t'] = exp_results[idx_exp][2]
-
-        # meas_level 2 return the shots
-        if m_lev == 2:
-
-            # convert the memory **array** into a n
-            # integer
-            # e.g. [1,0] -> 2
-            int_mem = memory.dot(np.power(2.0,
-                                          np.arange(memory.shape[1]))).astype(int)
-
-            # if the memory flag is set return each shot
-            if op_system.global_data['memory']:
-                hex_mem = [hex(val) for val in int_mem]
-                results['data']['memory'] = hex_mem
-
-            # Get hex counts dict
-            unique = np.unique(int_mem, return_counts=True)
-            hex_dict = {}
-            for kk in range(unique[0].shape[0]):
-                key = hex(unique[0][kk])
-                hex_dict[key] = unique[1][kk]
-
-            results['data']['counts'] = hex_dict
-
-        # meas_level 1 returns the <n>
-        elif m_lev == 1:
-
-            if m_ret == 'avg':
-
-                memory = [np.mean(memory, 0)]
-
-            # convert into the right [real, complex] pair form for json
-            # this should be cython?
-            results['data']['memory'] = []
-
-            for mem_shot in memory:
-                results['data']['memory'].append([])
-                for mem_slot in mem_shot:
-                    results['data']['memory'][-1].append(
-                        [np.real(mem_slot), np.imag(mem_slot)])
-
-            if m_ret == 'avg':
-                results['data']['memory'] = results['data']['memory'][0]
-
-        all_results.append(results)
-
-    if not op_system.use_cpp_ode_func:
-        _cython_build_cleanup(op_system.global_data['rhs_file_name'])
-
-    return all_results
 
 
 def unitary_evolution(exp, op_system):
